настройки раскиданы по вложенным страницам
