+++ lib/presentation/page/settings/widget/settings_card_widget.dart
@@ -8,18 +8,21 @@ class SettingsCardWidget extends StatelessWidget {
     super.key,
     this.title,
     this.subtitle,
+    this.actions = const [],
     this.padding = const .all(12),
     required this.child,
   });
 
   final String? title;
   final String? subtitle;
+  final List<Widget> actions;
   final EdgeInsets padding;
   final Widget child;
 
   @override
   Widget build(BuildContext context) {
     final theme = context.theme;
+    final needFirstRow = title != null || actions.isNotEmpty;
 
     return FlabrCard(
       margin: .zero,
@@ -28,12 +31,22 @@ class SettingsCardWidget extends StatelessWidget {
         crossAxisAlignment: .stretch,
         mainAxisSize: .min,
         children: [
-          if (title != null)
-            Text(
-              title!,
-              style: theme.textTheme.titleLarge!.copyWith(
-                fontWeight: .w400,
-              ),
+          if (needFirstRow)
+            Row(
+              children: [
+                if (title != null)
+                  Expanded(
+                    child: Text(
+                      title!,
+                      style: theme.textTheme.titleLarge!.copyWith(
+                        fontWeight: .w400,
+                      ),
+                    ),
+                  )
+                else
+                  const Spacer(),
+                ...actions,
+              ],
             ),
           if (subtitle != null)
             Text(subtitle!, style: theme.textTheme.bodyMedium),
